+++ src/Faker/Internal/Types.hs
@@ -173,6 +173,7 @@ data SourceData
   | Vehicle
   | VentureBros
   | Verbs
+  | Volleyball
   | Witcher
   | WorldCup
   | WorldOfWarcraft